Fixed bug where duplicate summary information was added.
I think this might be a better solution than what I had in previous commit. Basically, when we skip acts with JumpToAct, it adds that extra unneeded information in the summary so it can mark each act it skipped as done. We now restore the adventure summary twice in the checkpoint procedure so this is hopefully negated.